Fix form submission redirect not triggering component render
After a successful submission, redirecting to the movie list was
supposed to display the newly added movie on top of the list. But,
because of the way React Router redirecting works, it just navigated to
the movie index component without triggering a render, so the new movie
doesn't appear.

Also, trying to submit a form with validation errors was triggering a
redirect. I tried all kinds of different conditionals using Formik
values, but either the redirect never occurs or it fires on errors. Not
sure what that's about.